dev
pare down a lot more.   a full run was just 1:37 with a fair degree
of fast talking, no questions.   slowing the pace is essential
so do away with the "brief history" that doesnt really fit in
with a "tutorial" in any case, update/delete that are too brief
to be useful, "returning" thing that's not needed.